Refactored File Stucture
Moved all CSS files out of the components folders and into a css folder.
