+++ client/src/components/UIElements/FormElements/Input.js
@@ -1,7 +1,7 @@
 import React, { useReducer, useEffect } from "react";
 
 import { validate } from "../../Functions/validators";
-import "./Input.css";
+import "../../../css/Input.css";
 
 const inputReducer = (state, action) => {
   switch (action.type) {
